requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
tensorflow==2.12.0
numpy==1.23.5
pandas==2.0.1
apache-airflow==2.6.0
pytest==7.3.1
minio==7.1.14
opencv-python==4.8.1.78
python-dotenv==1.0.0
Pillow-11.2.1